move refactored tests
